Runner fixes 3.

+++ tools/draft_parser/src/ctrl_message_types.cpp
@@ -11,7 +11,7 @@ namespace quicr::messages {
 
     Bytes& operator<<(Bytes& buffer, const BytesSpan& bytes)
     {
-        buffer << bytes.size(); // length of byte span
+        buffer << static_cast<std::uint64_t>(bytes.size()); // length of byte span
         buffer.insert(buffer.end(), bytes.begin(), bytes.end());
         return buffer;
     }
